Use path.normalize to verify if either "." or ".." are incorrectly used
