+++ lib/endpoints/boot_params.js
@@ -7,8 +7,6 @@
 /*
  * Copyright 2020 Joyent, Inc.
  */
-const path = require('path');
-const util = require('util');
 
 const restify = require('restify');
 
@@ -17,67 +15,6 @@ const common = require('../common');
 const validation = require('../validation/endpoints');
 const ModelPlatform = require('../models/platform');
 
-/*
- * Expected boot modules format is as follows:
- *
- * {
- *      "path": "boot module path to be used by booter",
- *      "type": "Content type. Right now only base64 is supported",
- *      "content": "File contents, if necessary encoded as specified by type"
- * }
- *
- * In the future, support for new content types or additional members for
- * the boot_module objects could be provided, defaulting to base64.
- */
-function validateBootModules(boot_modules) {
-    var errs = [];
-    var i;
-    for (i = 0; i < boot_modules.length; i += 1) {
-        var bm = boot_modules[i];
-        bm.type = bm.type || 'base64';
-
-        if (bm.type !== 'base64') {
-            errs.push(util.format(
-                'Unsupported type %s for module \'%s\'', bm.type, bm.path));
-            continue;
-        }
-
-        if (!bm.path || !bm.content) {
-            errs.push(util.format(
-                'Unexpected format for boot_module: %j', bm));
-            continue;
-        }
-
-        if (/[.]+/.test(path.dirname(bm.path))) {
-            errs.push(util.format(
-                'Invalid path for boot_module: \'%s\' ' +
-                '(path cannot contain \'.\' or \'..\')', bm.path));
-            continue;
-        }
-
-        if (Buffer.byteLength(bm.content, 'base64') > 4 * 1024) {
-            errs.push(util.format(
-                'Module \'%s\' exceeds the maximum allowed size of 4KB',
-                bm.path));
-            continue;
-        }
-
-        if (Buffer.from(bm.content, 'base64').toString('base64') !==
-            bm.content) {
-            errs.push(util.format(
-                'Contents for module \'%s\' must be base64 encoded', bm.path));
-            continue;
-        }
-    }
-
-    if (errs.length) {
-        var err = new restify.InvalidArgumentError(
-            'Values provided for boot_modules contain errors: '
-            + errs.join(',\n'));
-        return err;
-    }
-    return null;
-}
 
 /**
  * Boot parameters are passed to compute nodes on boot. They enable one to pass
@@ -172,7 +109,7 @@ BootParams.setDefault = function handlerBootParamsSetDefault(req, res, next) {
 
 
     if (req.params.boot_modules) {
-        var bootmErr = validateBootModules(req.params.boot_modules);
+        var bootmErr = validation.validateBootModules(req.params.boot_modules);
         if (bootmErr) {
             res.send(bootmErr);
             next();
@@ -283,7 +220,7 @@ function handlerBootParamsUpdateDefault(req, res, next) {
     params.serial = req.params.serial;
 
     if (req.params.boot_modules) {
-        var bootmErr = validateBootModules(req.params.boot_modules);
+        var bootmErr = validation.validateBootModules(req.params.boot_modules);
         if (bootmErr) {
             res.send(bootmErr);
             next();
@@ -444,7 +381,7 @@ BootParams.setByUuid = function handlerBootParamsSetByUuid(req, res, next) {
     values.serial = req.params.serial;
 
     if (req.params.boot_modules) {
-        var bootmErr = validateBootModules(req.params.boot_modules);
+        var bootmErr = validation.validateBootModules(req.params.boot_modules);
         if (bootmErr) {
             res.send(bootmErr);
             next();
@@ -562,7 +499,7 @@ function handlerBootParamsUpdateByUuid(req, res, next) {
     params.serial = req.params.serial;
 
     if (req.params.boot_modules) {
-        var bootmErr = validateBootModules(req.params.boot_modules);
+        var bootmErr = validation.validateBootModules(req.params.boot_modules);
         if (bootmErr) {
             res.send(bootmErr);
             next();
