+++ lib/validation/endpoints.js
@@ -1,14 +1,17 @@
 /*
- * Copyright (c) 2019, Joyent, Inc. All rights reserved.
+ * Copyright 2020, Joyent, Inc. All rights reserved.
  *
  * HTTP endpoints validation routines.
  *
  */
 
-var sprintf = require('sprintf').sprintf;
-var VError = require('verror');
-var restify = require('restify');
-var restify_validator = require('restify-validator');
+const path = require('path');
+const util = require('util');
+
+const restify = require('restify');
+const restify_validator = require('restify-validator');
+const sprintf = require('sprintf').sprintf;
+const VError = require('verror');
 
 // Restcode = InvalidParameters
 
@@ -69,7 +72,7 @@ function ensureParamsValid(req, res, paramRules, opts) {
             rule = paramRules[paramName][ruleIdx];
             var ruleName;
 
-            if (global.toString.call(rule) == '[object String]') {
+            if (global.toString.call(rule) === '[object String]') {
                 ruleName = rule;
                 rule = [ruleName];
             } else if (Array.isArray(rule)) {
@@ -93,7 +96,6 @@ function ensureParamsValid(req, res, paramRules, opts) {
 
                     if (!params.hasOwnProperty(paramName) ||
                         params[paramName] === undefined) {
-
                         skip = true;
                         params[paramName] = rule[1];
                         break;
@@ -240,7 +242,7 @@ function ensureParamsValid(req, res, paramRules, opts) {
 function updateLimitedToInternalMetadata(req, res, name) {
     if (req.params.hasOwnProperty('update')) {
         var keys = Object.keys(req.params.update);
-        if (keys.length != 1 || keys[0] !== 'set_internal_metadata') {
+        if (keys.length !== 1 || keys[0] !== 'set_internal_metadata') {
             res.send(
                 500,
                 new restify.InvalidArgumentError(
@@ -252,8 +254,71 @@ function updateLimitedToInternalMetadata(req, res, name) {
     return null;
 }
 
+/*
+ * Expected boot modules format is as follows:
+ *
+ * {
+ *      "path": "boot module path to be used by booter",
+ *      "type": "Content type. Right now only base64 is supported",
+ *      "content": "File contents, if necessary encoded as specified by type"
+ * }
+ *
+ * In the future, support for new content types or additional members for
+ * the boot_module objects could be provided, defaulting to base64.
+ */
+function validateBootModules(boot_modules) {
+    var errs = [];
+    var i;
+    for (i = 0; i < boot_modules.length; i += 1) {
+        var bm = boot_modules[i];
+        bm.type = bm.type || 'base64';
+
+        if (bm.type !== 'base64') {
+            errs.push(util.format(
+                'Unsupported type %s for module \'%s\'', bm.type, bm.path));
+            continue;
+        }
+
+        if (!bm.path || !bm.content) {
+            errs.push(util.format(
+                'Unexpected format for boot_module: %j', bm));
+            continue;
+        }
+
+        if (path.normalize(bm.path) !== bm.path) {
+            errs.push(util.format(
+                'Invalid path for boot_module: \'%s\' ' +
+                '(path cannot contain \'.\' or \'..\')', bm.path));
+            continue;
+        }
+
+        if (Buffer.byteLength(bm.content, 'base64') > 4 * 1024) {
+            errs.push(util.format(
+                'Module \'%s\' exceeds the maximum allowed size of 4KB',
+                bm.path));
+            continue;
+        }
+
+        if (Buffer.from(bm.content, 'base64').toString('base64') !==
+            bm.content) {
+            errs.push(util.format(
+                'Contents for module \'%s\' must be base64 encoded', bm.path));
+            continue;
+        }
+    }
+
+    if (errs.length) {
+        var err = new restify.InvalidArgumentError(
+            'Values provided for boot_modules contain errors: '
+            + errs.join(',\n'));
+        return err;
+    }
+    return null;
+}
+
 exports.ensureParamsValid = ensureParamsValid;
 exports.formatValidationErrors = formatValidationErrors;
 exports.assert = {
     updateLimitedToInternalMetadata: updateLimitedToInternalMetadata
 };
+exports.validateBootModules = validateBootModules;
